refactor: adding model response types and moving run function to a new file.
